BugID:24820156: Add default config in aos_config.h and it can be convert to .config

+++ build/scripts/aos_check_config.py
@@ -10,21 +10,25 @@ BOARD_CONFIG_FILE = "board.config"
 def merge_new_config_with_default(dirname, boardname, appname):
     """ merge config file(.config) with app.config and board.config if existed, 
     and remove app.config and board.config """
-    if os.path.isdir(os.path.join(dirname, "core/rhino")):
+    if os.path.isdir(os.path.join(dirname, "core/rhino")) or os.path.isdir(os.path.join(dirname, "include/aos")):
         board_config = os.path.join(dirname, "platform/board", boardname, BOARD_CONFIG_FILE)
         appname = appname.replace(".", "/")
         app_config = os.path.join(dirname, "application/example", appname, APP_CONFIG_FILE)
+        in_appdir = False
     else:
         board_config = os.path.join(dirname, "board", boardname, BOARD_CONFIG_FILE)
         app_config = os.path.join(dirname, APP_CONFIG_FILE)
+        in_appdir = True
 
     config_file = os.path.join(dirname, DOT_CONFIG_FILE)
     if os.path.isfile(board_config):
         merge_config(config_file, board_config)
-        # os.remove(board_config)
+        if in_appdir:
+            os.remove(board_config)
     if os.path.isfile(app_config):
         merge_config(config_file, app_config)
-        os.remove(app_config)
+        if in_appdir:
+            os.remove(app_config)
     
 
 def check_appname(appname):
@@ -47,8 +51,9 @@ def check_appname(appname):
     return False
 
 def main():
-    """ check the appname and boardname between build string and .config file, 
-    and merge .config file with app.board and board.config """
+    """ after get default configuraiton by running app@board.config, check the appname 
+    and boardname between build string and .config file, and merge .config file with 
+    app.board and board.config """
     if len(sys.argv) < 3:
         print ("Usage: %s <config> <build_string>" % sys.argv[0])
         return 1
